+++ tests/unit/users/usersService.test.ts
@@ -35,7 +35,7 @@ describe('Users Service', () => {
       .post(`${AUTH_BASE}/oauth2/token`)
       .reply(200, { access_token: mockToken });
 
-    await sdk.authenticate({
+    await sdk.auth.authenticate({
       clientId: '',
       username: '',
       password: '',