+++ jest.config.js
@@ -2,15 +2,8 @@ module.exports = {
   verbose: true,
   testEnvironment: 'node',
   setupFiles: ['dotenv/config'],
-  moduleFileExtensions: [
-    'ts',
-    'tsx',
-    'js',
-    'json'
-  ],
-  transform: {
-    '\\.ts$': 'ts-jest'
-  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'json'],
+  preset: 'ts-jest',
   testRegex: '.*(\\.test|tests).*\\.(ts|js)$',
   // setupFiles: [
   //   "<rootDir>/tests/__helpers__/beforeEachSuite.ts"
@@ -19,10 +12,7 @@ module.exports = {
     '/node_modules/',
     '/tests/__helpers__/',
     '/tests/e2e/',
-    '/build/'
+    '/build/',
   ],
-  reporters: [
-    'default',
-    'jest-junit'
-  ]
-}
+  reporters: ['default', 'jest-junit'],
+};