+++ src/http/utils.test.ts
@@ -1,4 +1,11 @@
-import { recursiveMap } from './utils';
+import {
+  camelize,
+  camelizeKeys,
+  decamelize,
+  mapObjIndexed,
+  recursiveMap,
+  recursiveRenameKeys,
+} from './utils';
 
 describe('recursiveMap', () => {
   it('simple object mapEntriesRecursive', () => {
@@ -12,7 +19,7 @@ describe('recursiveMap', () => {
   it('complex object', () => {
     const result = recursiveMap((value, key) => {
       if (key.includes('stamp')) {
-        return new Date(value).toISOString();
+        return new Date(value);
       }
       return value;
     })({
@@ -35,8 +42,8 @@ describe('recursiveMap', () => {
         },
       ],
     });
-    expect(result.patient_enlistments[0].expiry_timestamp).toBe(
-      '2018-11-26T13:59:13.289Z'
+    expect(result.patient_enlistments[0].expiry_timestamp).toStrictEqual(
+      new Date('2018-11-26T13:59:13.289Z')
     );
   });
 
@@ -72,5 +79,112 @@ describe('recursiveMap', () => {
     expect(result[0].patient_enlistments[0].expiry_timestamp).toBe(
       '2018-11-26T13:59:13.289Z'
     );
+  });
+});
+
+describe('recursiveRenameKeys', () => {
+  it('simple object', () => {
+    const result = recursiveRenameKeys(value => `test_${value}`, {
+      test: 'value',
+    });
+
+    expect(Object.keys(result)).toStrictEqual(['test_test']);
+  });
+
+  it('complex object', () => {
+    const result = recursiveRenameKeys(
+      key => {
+        if (key === 'phone_number') {
+          return 'phone';
+        }
+        if (key === 'subObject') {
+          return 'objectSub';
+        }
+        if (key === 'patient_enlistments') {
+          return 'patientEnlistments';
+        }
+        if (key === 'expiry_timestamp') {
+          return 'expiry_ts';
+        }
+        return key;
+      },
+      {
+        id: '5bfbfc3146e0fb321rsa4b28',
+        phone_number: 9021324354,
+        subObject: { expiry_timestamp: 1543240753289 },
+        patient_enlistments: [
+          {
+            group_id: '5bfbfc3146e0fb321rsa4b28',
+            expiry_timestamp: 1543240753289,
+            expired: true,
+            creation_timestamp: 1543240753200,
+          },
+        ],
+      }
+    );
+    expect(result.patientEnlistments[0].expiry_ts).toBe(1543240753289);
+  });
+
+  it('complex array', () => {
+    const result = recursiveRenameKeys(
+      key => {
+        if (key === 'phone_number') {
+          return 'phone';
+        }
+        if (key === 'subObject') {
+          return 'objectSub';
+        }
+        if (key === 'patient_enlistments') {
+          return 'patientEnlistments';
+        }
+        if (key === 'expiry_timestamp') {
+          return 'expiry_ts';
+        }
+        return key;
+      },
+      [
+        {
+          id: '5bfbfc3146e0fb321rsa4b28',
+          phone_number: 9021324354,
+          subObject: { expiry_timestamp: 1543240753289 },
+          patient_enlistments: [
+            {
+              group_id: '5bfbfc3146e0fb321rsa4b28',
+              expiry_timestamp: 1543240753289,
+              expired: true,
+              creation_timestamp: 1543240753200,
+            },
+          ],
+        },
+      ]
+    );
+    expect(result[0].patientEnlistments[0].expiry_ts).toBe(1543240753289);
+  });
+});
+
+describe('camelize', () => {
+  it('easystring', () => {
+    const result = camelize('easy_string_to_test');
+    expect(result).toBe('easyStringToTest');
+  });
+  it('camelizeKeys', () => {
+    const result = camelizeKeys({
+      easy_string_to_test: 'test',
+    });
+    expect(result.easyStringToTest).toBeDefined();
+  });
+});
+
+describe('decamelize', () => {
+  it('easystring', () => {
+    const result = decamelize('easyStringTo_test');
+    expect(result).toBe('easy_string_to_test');
+  });
+});
+
+describe('mapObjIndexed', () => {
+  it('easy object', () => {
+    const result = mapObjIndexed(value => `-> ${value}`, { test: 'test' });
+    expect(result.test).toBe('-> test');
   });
 });