+++ src/services/data/index.ts
@@ -3,11 +3,17 @@ import httpClient from '../http-client';
 import infrastructureService from './infrastructureService';
 import schemasService from './schemasService';
 import transitionsService from './transitionsService';
+import indexesService from './indexesService';
+import statusesService from './statusesService';
+import propertiesService from './propertiesService';
 import { DATA_BASE } from '../../constants';
 
 export type DataService = ReturnType<typeof infrastructureService> &
   ReturnType<typeof schemasService> &
-  ReturnType<typeof transitionsService>;
+  ReturnType<typeof transitionsService> &
+  ReturnType<typeof indexesService> &
+  ReturnType<typeof statusesService> &
+  ReturnType<typeof propertiesService>;
 
 export default (
   http: HttpInstance,
@@ -20,10 +26,16 @@ export default (
   const infrastructureMethods = infrastructureService(client, http);
   const schemasMethods = schemasService(client, httpWithAuth);
   const transitionsMethods = transitionsService(client, httpWithAuth);
+  const indexesMethods = indexesService(client, httpWithAuth);
+  const statusesMethods = statusesService(client, httpWithAuth);
+  const propertiesMethods = propertiesService(client, httpWithAuth);
 
   return {
     ...infrastructureMethods,
     ...schemasMethods,
     ...transitionsMethods,
+    ...indexesMethods,
+    ...statusesMethods,
+    ...propertiesMethods,
   };
 };